requirements.txt
absl-py==2.0.0
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.8.0
aiohttp==3.8.6
aiosignal==1.3.1
ament-cmake-test==1.3.11
ament-copyright==0.12.12
ament-cppcheck==0.12.12
ament-cpplint==0.12.12
ament-flake8==0.12.12
ament-index-python==1.4.0
ament-lint==0.12.12
ament-lint-cmake==0.12.12
ament-package==0.14.0
ament-pep257==0.12.12
ament-uncrustify==0.12.12
ament-xmllint==0.12.12
angles==1.15.0
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
brotlipy==0.7.0
builtin-interfaces==1.2.1
cachetools==5.3.1
-e git+https://github.com/qintonguav/e2e-parking-carla.git@c3d03878284a75838f3cb1a9547f1a8c344a5aed#egg=carla&subdirectory=carla/PythonAPI/carla/dist/carla-0.9.11-py3.7-linux-x86_64
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cloudpickle==2.2.1
composition-interfaces==1.2.1
cryptography @ file:///croot/cryptography_1677533068310/work
cv-bridge==3.2.1
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
demo-nodes-py==0.20.5
diagnostic-msgs==4.8.0
distro==1.9.0
docstring_parser==0.16
domain-coordinator==0.10.0
efficientnet-pytorch==0.7.1
einops==0.6.1
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
geometry-msgs==4.8.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.23.3
google-auth-oauthlib==0.4.6
grpcio==1.59.0
gym==0.21.0
huggingface-hub==0.16.4
idna @ file:///croot/idna_1666125576474/work
image-geometry==3.2.1
importlib-metadata==4.13.0
interactive-markers==2.3.2
joblib==1.3.2
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
laser-geometry==2.4.0
launch==1.0.8
launch-ros==0.19.9
launch-testing==1.0.8
launch-testing-ros==0.19.9
launch-xml==1.0.8
launch-yaml==1.0.8
lifecycle-msgs==1.2.1
llvmlite==0.39.1
logging-demo==0.20.5
loguru==0.7.0
map-msgs==2.1.0
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1592846017414/work
message-filters==4.3.7
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
multidict==6.0.4
mypy-extensions==1.0.0
nav-msgs==4.8.0
numba==0.56.4
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.8.0.76
osrf-pycommon==2.1.6
packaging==23.2
pandas==1.3.5
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
Pillow==9.4.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyDeprecate==0.3.1
pygame==2.5.0
pynndescent==0.5.13
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
pyquaternion==0.9.9
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-qt-binding==1.1.2
pytorch-lightning==1.5.0
pytz==2025.2
PyYAML==6.0.1
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.16
rcutils==5.1.6
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2action==0.18.12
ros2bag==0.15.14
ros2cli==0.18.12
ros2component==0.18.12
ros2doctor==0.18.12
ros2interface==0.18.12
ros2launch==0.19.9
ros2lifecycle==0.18.12
ros2multicast==0.18.12
ros2node==0.18.12
ros2param==0.18.12
ros2pkg==0.18.12
ros2run==0.18.12
ros2service==0.18.12
ros2topic==0.18.12
rosbag2-interfaces==0.15.14
rosbag2-py==0.15.14
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
rsa==4.9
safetensors==0.4.0
scikit-learn==1.0.2
scikit-video==1.1.11
scipy==1.7.3
sensor-msgs==4.8.0
sensor-msgs-py==4.8.0
shape-msgs==4.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.2
sros2==0.10.6
stable-baselines3==1.6.2
statistics-msgs==1.2.1
std-msgs==4.8.0
std-srvs==4.8.0
stereo-msgs==4.8.0
teleop-twist-keyboard==2.4.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tf2-geometry-msgs==0.25.12
tf2-kdl==0.25.12
tf2-msgs==0.25.12
tf2-py==0.25.12
tf2-ros-py==0.25.12
tf2-tools==0.25.12
threadpoolctl==3.1.0
timm==0.9.7
topic-monitor==0.20.5
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.4
torchsummary==1.5.1
torchvision==0.14.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827244717/work
tqdm==4.66.1
trajectory-msgs==4.8.0
turtlesim==1.4.2
typed-argument-parser==1.8.0
typing-inspect==0.9.0
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
umap-learn==0.5.7
unique-identifier-msgs==2.2.1
urllib3 @ file:///croot/urllib3_1673575502006/work
visualization-msgs==4.8.0
Werkzeug==2.2.3
yarl==1.9.2
zipp==3.15.0
